+++ src/ts/Ads/Ads.ts
@@ -57,7 +57,6 @@ import { MRAID } from 'MRAID/MRAID';
 import { MRAIDView } from 'MRAID/Views/MRAIDView';
 import { PerformanceCampaign } from 'Performance/Models/PerformanceCampaign';
 import { Performance } from 'Performance/Performance';
-import { Promo } from 'Promo/Promo';
 import { VAST } from 'VAST/VAST';
 import { VPAID } from 'VPAID/VPAID';
 import { XPromoCampaign } from 'XPromo/Models/XPromoCampaign';
@@ -72,7 +71,6 @@ import { ARUtil } from 'AR/Utilities/ARUtil';
 import { PermissionsUtil, PermissionTypes } from 'Core/Utilities/Permissions';
 import { AbstractParserModule } from 'Ads/Modules/AbstractParserModule';
 import { MRAIDAdUnitParametersFactory } from 'MRAID/AdUnits/MRAIDAdUnitParametersFactory';
-import { PromoCampaign } from 'Promo/Models/PromoCampaign';
 import { PrivacyUnit } from 'Ads/AdUnits/PrivacyUnit';
 import { IStore } from 'Store/IStore';
 import { Store } from 'Store/Store';
@@ -85,7 +83,7 @@ import { Analytics } from 'Analytics/Analytics';
 import { PrivacySDK } from 'Privacy/PrivacySDK';
 import { PrivacyParser } from 'Privacy/Parsers/PrivacyParser';
 import { Promises } from 'Core/Utilities/Promises';
-import { LoadExperiment, LoadRefreshV4, MediationCacheModeAllowedTest, AuctionXHR, AuctionV6Test, LoadV5, BaseLineLoadV5 } from 'Core/Models/ABGroup';
+import { LoadExperiment, LoadRefreshV4, MediationCacheModeAllowedTest, AuctionXHR, AuctionV6Test, LoadV5, BaseLineLoadV5, TemporaryCacheModeAllowedTest } from 'Core/Models/ABGroup';
 import { PerPlacementLoadManagerV4 } from 'Ads/Managers/PerPlacementLoadManagerV4';
 import { PrivacyMetrics } from 'Privacy/PrivacyMetrics';
 import { PrivacySDKUnit } from 'Ads/AdUnits/PrivacySDKUnit';
@@ -96,7 +94,6 @@ import { MediationLoadTrackingManager, MediationExperimentType } from 'Ads/Manag
 import { CachedUserSummary } from 'Privacy/CachedUserSummary';
 import { createMeasurementsInstance } from 'Core/Utilities/TimeMeasurements';
 import { XHRequest } from 'Core/Utilities/XHRequest';
-import { NofillImmediatelyManager } from 'Ads/Managers/NofillImmediatelyManager';
 import { LegacyCampaignManager } from 'Ads/Managers/LegacyCampaignManager';
 import { AutomatedExperimentManager } from 'Ads/Managers/AutomatedExperimentManager';
 import { CampaignAssetInfo } from 'Ads/Utilities/CampaignAssetInfo';
@@ -124,7 +121,6 @@ export class Ads implements IAds {
     public CampaignManager: CampaignManager;
     public RefreshManager: RefreshManager;
     public MediationLoadTrackingManager: MediationLoadTrackingManager;
-    public NofillImmediatelyManager: NofillImmediatelyManager;
 
     private static _forcedConsentUnit: boolean = false;
 
@@ -133,7 +129,6 @@ export class Ads implements IAds {
     private _showingPrivacy: boolean = false;
     private _loadApiEnabled: boolean = false;
     private _webViewEnabledLoad: boolean = false;
-    private _nofillImmediately: boolean = false;
     private _forceLoadV5: boolean = false;
     private _mediationName: string;
     private _core: ICore;
@@ -194,28 +189,15 @@ export class Ads implements IAds {
     }
 
     public initialize(): Promise<void> {
-        const measurements = createMeasurementsInstance(InitializationMetric.WebviewInitializationPhases);
+        let measurements = createMeasurementsInstance(InitializationMetric.WebviewInitializationPhases);
         return Promise.resolve().then(() => {
             SdkStats.setInitTimestamp();
             GameSessionCounters.init();
-            Diagnostics.setAbGroup(this._core.Config.getAbGroup());
             return this.setupTestEnvironment();
         }).then(() => {
             this.setupLoadApiEnabled();
             measurements.overrideTag('wel', `${this._webViewEnabledLoad}`);
-
-            let promise = Promise.resolve();
-
-            this._nofillImmediately = CustomFeatures.isNofillImmediatelyGame(this._core.ClientInfo.getGameId()) && !!(performance && performance.now);
-            if (this._nofillImmediately) {
-                promise = this.setupMediationTrackingManager().then(() => {
-                    this.NofillImmediatelyManager = new NofillImmediatelyManager(this.Api.LoadApi, this.Api.Listener, this.Api.Placement, this.Config.getPlacementIds());
-                    return this._core.Api.Sdk.initComplete();
-                });
-            }
-            return promise;
         }).then(() => {
-            measurements.measure('setup_environment');
             return this.Analytics.initialize();
         }).then((gameSessionId: number) => {
             this.SessionManager.setGameSessionId(gameSessionId);
@@ -229,15 +211,12 @@ export class Ads implements IAds {
 
             PrivacyDataRequestHelper.init(this._core);
         }).then(() => {
-            measurements.measure('privacy_init');
             return this.setupMediationTrackingManager();
         }).then(() => {
-            measurements.measure('mediation_tracking_init');
             return this.PrivacyManager.getConsentAndUpdateConfiguration().catch(() => {
                 // do nothing since it's normal to have undefined developer consent
             });
         }).then(() => {
-            measurements.measure('consent_update');
             const defaultPlacement = this.Config.getDefaultPlacement();
             this.Api.Placement.setDefaultPlacement(defaultPlacement.getId());
 
@@ -246,16 +225,8 @@ export class Ads implements IAds {
                 this.AssetManager.setCacheDiagnostics(true);
             }
 
-            const promo = new Promo(this._core, this, this._core.Purchasing);
-            const promoContentTypeHandlerMap = promo.getContentTypeHandlerMap();
-            for (const contentType in promoContentTypeHandlerMap) {
-                if (promoContentTypeHandlerMap.hasOwnProperty(contentType)) {
-                    this.ContentTypeHandlerManager.addHandler(contentType, promoContentTypeHandlerMap[contentType]);
-                }
-            }
-
             this.BannerModule = new BannerModule(this._core, this);
-            this.Monetization = new Monetization(this._core, this, promo, this._core.Purchasing);
+            this.Monetization = new Monetization(this._core, this);
             this.AR = new AR(this._core);
 
             if (this.SessionManager.getGameSessionId() % 1000 === 0) {
@@ -298,8 +269,6 @@ export class Ads implements IAds {
             this.configureRefreshManager();
             SdkStats.initialize(this._core.Api, this._core.RequestManager, this._core.Config, this.Config, this.SessionManager, this.CampaignManager, this._core.MetaDataManager, this._core.ClientInfo, this._core.CacheManager);
 
-            promo.initialize();
-
             this.Monetization.Api.Listener.isMonetizationEnabled().then((enabled) => {
                 if (enabled) {
                     this.Monetization.initialize();
@@ -307,29 +276,23 @@ export class Ads implements IAds {
             });
 
         }).then(() => {
-            measurements.measure('managers_init');
             return this._core.Api.Sdk.initComplete();
         }).then(() => {
-            measurements.measure('init_complete_to_native');
             if (this.MediationLoadTrackingManager) {
                 this.MediationLoadTrackingManager.setInitComplete();
-            }
-
-            if (this.NofillImmediatelyManager) {
-                this.NofillImmediatelyManager.setInitComplete();
             }
 
             if (this.PrivacyManager.isPrivacySDKTestActive()) {
                 CachedUserSummary.fetch(this.PrivacyManager);
             }
 
+            measurements = createMeasurementsInstance(InitializationMetric.WebviewInitializationPhases);
+
             return Promises.voidResult(this.RefreshManager.initialize());
         }).then(() => {
             measurements.measure('request_on_init');
             return Promises.voidResult(this.SessionManager.sendUnsentSessions());
         }).then(() => {
-            measurements.measure('ads_ready');
-
             if (performance && performance.now) {
                 const webviewInitTime = performance.now();
                 SDKMetrics.reportTimingEventWithTags(InitializationMetric.WebviewInitialization, webviewInitTime, {
@@ -419,11 +382,11 @@ export class Ads implements IAds {
                 if (mediation && mediation.getName() && performance && performance.now) {
 
                     let experimentType = MediationExperimentType.None;
-                    if (CustomFeatures.isCacheModeAllowedTestGame(this._core.ClientInfo.getGameId()) && MediationCacheModeAllowedTest.isValid(this._core.Config.getAbGroup())) {
+                    const whitelistedCMATest = CustomFeatures.isCacheModeAllowedTestGame(this._core.ClientInfo.getGameId()) && MediationCacheModeAllowedTest.isValid(this._core.Config.getAbGroup());
+                    const temporaryCMATest = TemporaryCacheModeAllowedTest.isValid(this._core.Config.getAbGroup());
+                    if (whitelistedCMATest || temporaryCMATest) {
                         this.Config.set('cacheMode', CacheMode.ALLOWED);
                         experimentType = MediationExperimentType.CacheModeAllowed;
-                    } else if (this._nofillImmediately) {
-                        experimentType = MediationExperimentType.NofillImmediately;
                     } else if (this._core.NativeBridge.getPlatform() === Platform.ANDROID && AuctionXHR.isValid(this._core.Config.getAbGroup())) {
                         if (XHRequest.isAvailable()) {
                             experimentType = MediationExperimentType.AuctionXHR;
@@ -552,12 +515,6 @@ export class Ads implements IAds {
 
         SdkStats.sendShowEvent(placementId);
 
-        if (campaign instanceof PromoCampaign && campaign.getRequiredAssets().length === 0) {
-            this.showError(false, placementId, 'No creatives found for promo campaign');
-            SDKMetrics.reportMetricEvent(ErrorMetric.PromoWithoutCreatives);
-            return;
-        }
-
         if (campaign.isExpired()) {
             this.showError(true, placementId, 'Campaign has expired');
             this.RefreshManager.refresh();
@@ -669,7 +626,7 @@ export class Ads implements IAds {
                 // AdRequestManager is valid only if Load V5 is enabled
                 if (this.AdRequestManager) {
                     SDKMetrics.reportMetricEvent(LoadV5Metrics.Show);
-                } else if (this._webViewEnabledLoad && BaseLineLoadV5.isValid(this._core.Config.getAbGroup()) && CustomFeatures.isLoadV5Game(this._core.ClientInfo.getGameId())) {
+                } else if (this._loadApiEnabled && BaseLineLoadV5.isValid(this._core.Config.getAbGroup()) && CustomFeatures.isLoadV5Game(this._core.ClientInfo.getGameId())) {
                     SDKMetrics.reportMetricEvent(LoadV5Metrics.Show);
                 }