+++ src/ts/Ads/Models/AutomatedExperimentsList.ts
@@ -20,6 +20,11 @@ export enum AutomatedExperimentsCategories {
 }
 
 export const ButtonExperimentDeclaration: IExperimentDeclaration = {
+  scheme: {
+      DARK: 'dark',
+      LIGHT: 'light',
+      COLORMATCHING: 'color_matching'
+  },
   animation: {
       STATIC: 'static',
       HEARTBEATING: 'heartbeating',
@@ -35,11 +40,12 @@ export const ButtonExperimentDeclaration: IExperimentDeclaration = {
 };
 
 export const ButtonAnimationsExperiment = new AutomatedExperiment({
-    name: 'wbvw-endcard-v1',
+    name: 'wbvw-endcard-v2',
     actions: ButtonExperimentDeclaration,
     defaultActions: {
-        color: ButtonExperimentDeclaration.color.BLUE,
-        animation: ButtonExperimentDeclaration.animation.BOUNCING
+        scheme: ButtonExperimentDeclaration.scheme.LIGHT,
+        animation: ButtonExperimentDeclaration.animation.BOUNCING,
+        color: ButtonExperimentDeclaration.color.BLUE
     },
     cacheDisabled: true
 });