+++ src/ts/Core/Models/ABGroup.ts
@@ -77,13 +77,13 @@ export const FakeReverseABTest = new ReverseABTest(16, 17);
 
 // Active AB Tests
 export const MediationCacheModeAllowedTest = new ReverseABTest(5, 6);
+export const TemporaryCacheModeAllowedTest = new ABTest(15);
 export const MabDisabledABTest = new ZyngaFilteredABTest(7);
 export const LoadExperiment = new ABTest(13);
 export const LoadRefreshV4 = new ABTest(14);
 export const PrivacySDKTest = new ABTest(16);
-export const DiagnosticsRampDown = new ABTest(15);
 export const AuctionXHR = new DisabledABTest();
-export const AuctionV6Test = new DisabledABTest(18);
+export const AuctionV6Test = new ABTest(18);
 export const LoadV5 = new DisabledABTest(19);
 export const BaseLineLoadV5 = new ABTest(19);